==> 1_Predictors.Rmd <==
---
title: "Build predictor stack and extract their values to sample points"
output: html_notebook
---

# Intro
Inputs: sample_info
Outputs: the predictor brick ("pred"), "e" data frame (values from raster stack extracted to all sample points)

# Libraries
```{r}
library(raster)
library(rgdal)
library(rgeos)
#install_github("skgrange/threadr")
library(threadr)
library(RCurl)
```

# Read in NGU Data
make "pred" object
```{r}
files <- list_files_ftp(
  ftp_url,
  credentials = "havforsk:Havf8776",
  sleep = NA,
  sort = FALSE,
  verbose = FALSE
)
x <- stack()
for (i in (seq_along(files)[-length(seq_along(files))])) {
  print(i)
  tmpraster <- raster(paste0(ftp_filepath,
                            paste(unlist(strsplit(files[i], split="all", fixed=TRUE))[2], sep=""))
  )
  names(tmpraster) <- substr(paste(unlist(strsplit(files[i], split="all", fixed=TRUE))[2], sep=""),
              2,
              nchar(paste(unlist(strsplit(files[i], split="all", fixed=TRUE))[2], sep=""))-4)
  x <- stack( x , tmpraster)
  
}
pred <- x
pred <- dropLayer(pred, which(names(pred)%in%gsub(".tif", "", list.files("G:\\GIS_Data\\EnvironmentalLayersAdditional"))))
rm(x, tmpraster)
pred
```

# Extract values to points
## Get sample info
### If Marvid runing use this:
```{python}
import requests

API_URL = "http://marvid-staging.hi.no:8092/query/video_lines"
OUTPUL_FILEPATH = "C:\\Users\\genoveva\\Downloads\\marvid_sections.csv"
DELIMITER = ","
CSV_HEADER = DELIMITER.join(["video line", "section", "mean latitude", "mean longitude"])

api_response = requests.get(API_URL).json() 
output_strings = [CSV_HEADER] 
for vl_number in sorted(api_response.keys()):
    vl = api_response[vl_number]
    for section in sorted(vl["sections"]):
        mean_lat = str(vl["sections"][section]["mean_lat"])
        mean_lon = str(vl["sections"][section]["mean_lon"])
        output_strings.append(DELIMITER.join([vl_number, section, mean_lat, mean_lon]))

with open(OUTPUL_FILEPATH, "w") as output_file:
    output_file.write("\n".join(output_strings))n
```

### If not, use this:
```{r}
sample_info <- read.csv(file.path(Data, "sample_info.csv"))
sample_info <- sample_info %>% rename(video.line=VL, section=SampID2, mean.latitude=y_coord, mean.longitude=x_coord) %>%
  dplyr::select(video.line, section, mean.latitude, mean.longitude, SampID)
# sample_info <- sample_info %>% makesampid # only needed when we switch to using marvid sections
```

## Extract
```{r}
e <- raster::extract(pred,dplyr::select(sample_info, mean.longitude, mean.latitude), sp=TRUE)
e <- cbind(e,sample_info)
```

# Save outputs
```{r}
write.csv(e, "G:\\R_projects\\VH_MarVid\\Intermediate_objects\\pred_data\\e.csv" )

# save whole workspace (writing pred to tif will delete layer names, which is a pain...)
save.image(file=file.path(savepath, "pred_data/1_Predictors_out.RData"))

# But also write to tiff, who cares
writeRaster(pred, filename=file.path(savepath,"pred_data/pred.tif"), options="INTERLEAVE=BAND", overwrite=TRUE)

# write to spatial pixels
saveRDS(as(pred, "SpatialPixelsDataFrame"), file.path(savepath,"pred_data/pred.rds"))
```
