==> 2B_Species_different_niche.Rmd <==
---
title: "Calculate niche overlap between indicators of same VME"
output: html_notebook
---
# Libraries
```{r}
library(dplyr)
library(tidyr)
library(caret)
library(nicheROVER)
library(corrplot)
library(ecospat)
```

# Inputs
```{r}
vme <-  "Hard bottom gorgonian gardens"
sp = taxonary %>% filter(VME_Burgos_etal_2020==vme) %>% dplyr::select(Reference_List) %>% pull
#sp = taxonary %>% filter(VME_BuhlMortensen_etal_2023==vme) %>% dplyr::select(Reference_List) %>% pull
#e <- read.csv(file.path(savepath, "e.csv"))
# data, from fake Marvid report script

# Parameters
perc <- 2 # 1 is 1st percentile, 2 is 5th percentile
nsamples <- 10000
toler <- 10
sp
```

# Filter data
```{r}
# obtain only rows of interest
data_extract <- data %>% filter(taxon%in%sp) %>% left_join(e,by = join_by(SampID == section))
#hist(log(data_extract$density[which(data_extract$taxon=="Antho dichotoma")]))
```

# Define the density threshold for each species
This intends to separate thriving populations from occasional occurrences. Needs to set the desired quantile. Use 1 and 5 %!!!
```{r}
# create empty threshold vector
threshold <- vector(mode = "numeric",
                   length = length(sp))

for (i in 1:length(sp)){
  threshold[i] <- (data %>% filter(taxon==sp[i]) %>% pull(density) %>% sort %>%  quantile(., c(.01, .05)))[perc]
}

thresholdt <- data.frame(sp=sp, threshold=as.numeric(threshold))
thresholdt
```

## Further filter the data
and select columns
```{r}
data_extract_merged <-merge(data_extract,thresholdt,by.x="taxon",by.y="sp",all.x=TRUE)
data_extract_merged_reduced <- data_extract_merged %>% filter(density>threshold) %>%
  dplyr::select(taxon,          
         bathy,                        
         BO22_carbonphytoltmax_bdmean,
         BO22_carbonphytoltmax_ss,
         BO22_carbonphytoltmin_bdmean,
         BO22_carbonphytoltmin_ss,  
         BO22_carbonphytomean_bdmean,
         BO22_carbonphytomean_ss,
         BO22_carbonphytorange_bdmean,
         BO22_carbonphytorange_ss,
         BO22_chloltmax_bdmean,
         BO22_chloltmax_ss,    
         BO22_chloltmin_bdmean,       
         BO22_chloltmin_ss,  
         BO22_chlomean_bdmean,    
         BO22_chlomean_ss,  
         BO22_chlorange_bdmean,       
         BO22_chlorange_ss,        
         BO22_dissoxltmax_bdmean,   
         BO22_dissoxltmin_bdmean,  
         BO22_dissoxmean_bdmean,      
         BO22_dissoxrange_bdmean,   
         BO22_icecoverltmax_ss,     
         BO22_icecoverltmin_ss,     
         BO22_icecovermean_ss,        
         BO22_icecoverrange_ss,     
         BO22_icethickltmax_ss,     
         BO22_icethickltmin_ss,    
         BO22_icethickmean_ss,        
         BO22_icethickrange_ss,    
         BO22_ironltmax_bdmean,     
         BO22_ironltmin_bdmean,     
         BO22_ironmean_bdmean,        
         BO22_ironrange_bdmean,     
         BO22_lightbotltmax_bdmean,  
         BO22_lightbotltmin_bdmean,  
         BO22_lightbotmean_bdmean,    
         BO22_lightbotrange_bdmean,  
         BO22_nitrateltmax_bdmean,   
         BO22_nitrateltmin_bdmean, 
         BO22_nitratemean_bdmean,     
         BO22_nitraterange_bdmean,    
         BO22_phosphateltmax_bdmean, 
         BO22_phosphateltmin_bdmean,  
         BO22_phosphatemean_bdmean,   
         BO22_phosphaterange_bdmean, 
         BO22_ppltmax_bdmean,       
         BO22_ppltmax_ss,            
         BO22_ppltmin_bdmean,         
         BO22_ppltmin_ss,            
         BO22_ppmean_bdmean,        
         BO22_ppmean_ss,           
         BO22_pprange_bdmean,         
         BO22_pprange_ss,           
         BO22_silicateltmax_bdmean,   
         BO22_silicateltmin_bdmean,   
         BO22_silicatemean_bdmean,    
         BO22_silicaterange_bdmean,   
         CDirmax_Robinson,         
         CDirmean_Robinson,        
         CDirmin_Robinson,            
         CDirsd_Robinson,          
         cobB,                    
         CSpdmax_Robinson,        
         CSpdmean_Robinson,           
         CSpdmin_Robinson,         
         CSpdsd_Robinson,          
         diffME3,                   
         diffME9,                     
         gravel,                      
         landscape,                  
         MLDmax_Robinson,             
         MLDmean_Robinson,            
         MLDmin_Robinson,             
         MLDsd_Robinson,             
         msr1_mag,                    
         msr5_mag,                    
         MS_biogeo05_dist_shore_5m,   
         mud,                        
         rock,                        
         salt_max,                    
         salt_mean,                  
         salt_min,                  
         salt_std,                    
         sand,                        
         sedclass,                   
         seddan,                    
         sedmil,                    
         slope3,                      
         slope9,                    
         Smax_Robinson,              
         Smean_Robinson,             
         Smin_Robinson,               
         spd_max,                     
         spd_mean,                   
         spd_min,                  
         spd_std,                     
         Ssd_Robinson,              
         temp_max,                 
         temp_mean,                
         temp_min,                    
         temp_std,               
         Tmax_Robinson,           
         Tmean_Robinson,       
         Tmin_Robinson,               
         Tsd_Robinson,   
         Umax_Robinson,    
         Umean_Robinson,   
         Umin_Robinson,               
         Usd_Robinson,   
         u_bott_mean,     
         Vmax_Robinson,    
         Vmean_Robinson,              
         Vmin_Robinson,    
         Vsd_Robinson,      
         v_bott_mean    )

data_extract_merged_reduced <- data_extract_merged_reduced %>% filter(complete.cases(.))
data_extract_merged_reduced
```

# Remove variables
```{r}
df <- e %>% dplyr::select(-c(1,119:123)) %>% # sample ID and other metadata
  filter(complete.cases(.))

# identify zero variance predictors
zv <- apply(df, 2, function(x) length(unique(x)) == 1)
dfr <- df[, !zv]

# identify highly colinear
correlationMatrix <- cor(dfr[,1:dim(dfr)[2]],use="complete.obs")
namesToDrop <- findCorrelation(cor(dfr, use="pairwise.complete.obs"), cutoff = 0.8, names=TRUE)

namesToDrop <- c(namesToDrop, names(which(zv)))

data_extract_merged_reduced_new <- data_extract_merged_reduced %>%
  dplyr::select(-na.omit(match(namesToDrop, colnames(data_extract_merged_reduced))))
data_extract_merged_reduced_new
```


## Calculate niche overlap
```{r}
d<-data_extract_merged_reduced_new

# Identify species with too few observations
few <- names(which(table(d$taxon)<dim(d)[2]))

d <- d %>% filter(!taxon%in%few)

d <- d %>% dplyr::select(-c(  landscape, sedclass, seddan, sedmil, #factor variables
                       BO22_icecovermean_ss,
                       BO22_lightbotltmax_bdmean,
                       BO22_ppltmin_bdmean,
                       BO22_ppltmin_ss,
                       gravel)) 

d$taxon<- as.factor(as.character(d$taxon))
```

# Visualize it
```{r}
plot_fun <- function(x) {
  ggplot(data = d, aes(x = .data[[x]], color = taxon, show.legend = FALSE)) +
    geom_density()
}

cols <- names(d)[!names(d) %in% c("taxon")]
names(cols) <- cols

gg<-lapply(cols, plot_fun)
gg
```

# Calculate niche overlap
And generate desired output
```{r}
d.par <- tapply(1:nrow(d), d$taxon,
                   function(ii) niw.post(nsamples = nsamples, X = d[ii,2: dim(d)[2]]))

# Overlap calculation.  use nsamples = nprob = 10000 (1e4) for higher accuracy.
# the variable over.stat can be supplied directly to the overlap.plot function

over.stat <- overlap(d.par, nreps = nsamples, nprob = 1e3, alpha = c(.95, 0.99))

#The mean overlap metrics calculated across iterations for both niche 
#region sizes (alpha = .95 and alpha = .99) can be calculated and displayed in an array.
over.mean <- apply(over.stat, c(1:2,4), mean)*100
coex<-round(over.mean, 2)[,,1] # for alpha=95%
coex
```

# result
```{r}
result <- cbind(rowMeans(data.frame(coex), na.rm=TRUE), colMeans(data.frame(coex), na.rm=TRUE)) %>%
  data.frame %>%
  arrange(X2)
result
```
# find the most "indicative" indicators
Those with large values in second column have the most species within them (umbrella value?).
Find species with largest "umbrella Value", take out and re-calculate the means. Count how many stations remain.
New idea: the larger the value in the second column, the more generalist the species (?). Model performace deacreases when eliminating from the group species with low value in X2 (e.g., Craniella polyura, Phakellia/Axinella)
```{r}
rogue <- filter(result, X2<toler) %>% row.names(.)
rogue
```
# plot the co-occurrence matrix
```{r}
data_extract_merged_binned <-  data_extract_merged %>% mutate(presence=case_when(density>threshold ~ 1,
                                                                                 density<=threshold ~ 0)) %>%
  dplyr::select(taxon,          
         SampID,
         presence)
         
         
matrix <- pivot_wider(data_extract_merged_binned, id_cols=SampID, values_from = presence, names_from=taxon, values_fill = 0)
matrix<-matrix %>% filter(!is.na(SampID))
M<-ecospat.co_occurrences (data=matrix)
corrplot(M, order="AOE")
```
