==> 4B_Model_select_quantregForest.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

inputs: v; resp_spat (whole video line level), pred (as spatial pixels) (the latter two only needed if generating spatial variables)
Workflow: feature selection, parameter fine tuning, train final model with and without weights
outputs: m1, m2, m3, vecv, and ccc

# Libraries etc
```{r}
library(ranger)
library(groupdata2)
library(quantregForest)
library(tidyverse)
library(rgdal)
library(caret)
library(spm)
library(DescTools)
library(doParallel)
#library(doMC) # unix only
library(yardstick)
library(BAMMtools)

quantiles = c((1-.682)/2, 0.5, 1-(1-.682)/2)
remove=""
```

# Inputs
```{r}
# get v (read)
v <- read.csv(file.path(savepath, paste(vme, "minus", paste(substr(remove, 1,7), collapse=", "), "v.csv", collapse=" ")))
# str(v)

v <- v %>% 
      mutate_at(vars(  landscape, sedclass, seddan, sedmil    ), factor)

v <- v %>% filter(complete.cases(.))
dim(v)
```

# More inputs: resp spat
only needed if generating spatial variables!!
```{r}
resp_spat_vl <- readRDS(file.path(savepath, paste(vme,"minus", paste(substr(remove, 1,7), collapse=", "),"resp_Spat_vl.rds")))
#resp_spat <- readRDS(file.path(savepath, paste(vme,"minus", paste(substr(remove, 1,7), collapse=", "), "resp_spat.rds")))
```

# More inputs: pred
```{r}
#load(file.path(savepath,"1_Predictors_out.RData"))
pred.pix <- readRDS(file.path(savepath,"pred_data/pred.rds"))
pred.pix1 <- subset(pred.pix, complete.cases(pred.pix@data))
pred.pix1@data <- pred.pix1@data %>% mutate_at(vars(  landscape, sedclass, seddan, sedmil    ), factor)
```

# select environmental variables
```{r}
cl <- makeCluster(detectCores(), type='PSOCK')
registerDoParallel(cl)
control <- rfeControl(functions=rfFuncs, method="cv", number=10)
# run the RFE algorithm
a=which(colnames(v)=="bathy")
b=which(colnames(v)=="v_bott_mean")
c=which(colnames(v)=="vmeind_density")
results <- rfe(v[,a:b], v[,c], sizes=c(1:(b-a)), rfeControl=control)
# chosen features
selvar <- predictors(results)
selvar
```

# add spatial variables
```{r}
resp.dist0 <-buffer.dist(resp_spat_vl["vmeind_density"], 
                                 pred.pix[1], as.factor(1:nrow(resp_spat_vl))) 

ov.resp = over(resp_spat["vmeind_density"], resp.dist0)

ov.resp <- over(resp_spat["vmeind_density"], pred.pix[1:2])
sw.rm = do.call(cbind, list(resp_spat@data["vmeind_density"], ov.resp
                            , ov.resp
                            ))

v <- v %>% left_join(sw.rm) %>% filter(complete.cases(.))
```

# Set model formula
```{r}
all_names <- paste(#c(names(resp.dist0),
                     selvar, collapse="+")
fmla0 <- as.formula(paste("vmeind_density ~ ", 
                           all_names))
fmla0
```

# fine tune mtry parameter
```{r}
# configure multicore
#registerDoMC(cores=12)
cl <- makeCluster(detectCores(), type='PSOCK')
registerDoParallel(cl)

control <- trainControl(method="repeatedcv", number=10, repeats=3, search="grid")
set.seed(1)
tunegrid <- expand.grid(.mtry=c(1:length(selvar)))
rf_gridsearch <- train(fmla0, data=v, method="qrf", metric="Rsquared", tuneGrid=tunegrid, trControl=control)
print(rf_gridsearch)
mtry=rf_gridsearch$bestTune
```

# train new model
without case weights
```{r}
m1 <- ranger(formula = fmla0,
             data = v,
             mtry=pull(mtry), 
             min.node.size=2,
             sample.fraction=0.9930754, 
             num.trees=250,
             importance = "impurity",
             seed=1,
             quantreg=TRUE,
             respect.unordered.factors = "partition"
             )
m1
```

with case weights
```{r}
m2 <- ranger(formula = fmla0,
             data = v,
             mtry=pull(mtry), 
             min.node.size=2,
             sample.fraction=0.9930754, 
             num.trees=250,
             importance = "impurity",
             seed=1,
             quantreg=TRUE,
             respect.unordered.factors = "partition",
             case.weights = as.integer(1/(v$sd_dens^2))
             )
m2
```

# Split data for model testing
split the data into two partitions where (1) video lines are kept together and (2) the testing data is biased towards intermediate to high-density observations
```{r}
p=0.05# proportion of the core dataset will be used for testing the model (fringe dataset will be 50/50)

split <- stratisplit2(v=v, pr=p)

v_test <- split[[1]]
v_train <- split[[2]]

rbind(hist(log(v_train$vmeind_density)), hist(log(v_test$vmeind_density)))
```

# test model
```{r}
m3 <- ranger(formula = fmla0,
             data = v_train,
             mtry=pull(mtry), 
             min.node.size=2,
             sample.fraction=0.9930754, 
             num.trees=250,
             importance = "impurity",
             seed=1,
             quantreg=TRUE,
             respect.unordered.factors = "partition"#,
             #case.weights = as.integer(1/(v$sd_dens^2))
             )
m3
```

# variance explained by cross-validation (from spm library)
```{r}
vme.qrf2 <- predict(m3,  v_test, 
                     type="quantiles", quantiles=quantiles)$predictions
## now more computational...
v_test$vme.qrf2pr = vme.qrf2[,2]
## s.d. of the prediction error:
v_test$vme.qrf2var = (vme.qrf2[,3]-vme.qrf2[,1])/2

predicted <- v_test$vme.qrf2pr
observed <- v_test %>%
  dplyr::select(vmeind_density)

vecv <- vecv(pull(observed),pull(data.frame(unlist(predicted)))) # variance explained by cross-validation, conservative estimate to report
vecv
```

# concordance correlation coefficient
```{r}
ccc <- ccc_vec(pull(observed),predicted)
ccc
```
# Save outputs
```{r}

saveRDS(m3, file = file.path(outpath,paste(vme,"minus", paste(substr(remove, 1,7), collapse=", "), "model.rda" )))
write.csv(data.frame(SampID=c(as.character(v_test$SampID), as.character(v_train$SampID)), set=c(rep("test",length(v_test$SampID)),
                                                                                                rep("train", length(v_train$SampID)))),
          file.path(outpath, paste(vme,"minus", paste(substr(remove, 1,7), collapse=", "), "data split.csv")),
          row.names = FALSE)
```
