==> 5_Model_predict.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Intro

This script can be run as soon as environmental layers for new areas are available, i.e. without the MarVid data yet ready. Requires running "1_Predictors" first, and for the names not to change

# Libraries
```{r}
library(raster)
## color legend:
leg = c("#0000ff", "#0028d7", "#0050af", "#007986", "#00a15e", "#00ca35", 
        "#00f20d", "#1aff00", "#43ff00", "#6bff00", "#94ff00", "#bcff00", 
        "#e5ff00", "#fff200", "#ffca00", "#ffa100", "#ff7900", "#ff5000", "#ff2800", "#ff0000")
```

# inputs
```{r}
#if m3 performance was "similar" to m2, use m2 (same parameters, all data)
#m <- m1
vme <- "Cauliflower coral fields"
remove <- "nothing"
name <- "m2"

v <- read.csv(file.path(savepath, paste(vme, "minus", paste(substr(remove, 1,7), collapse=", "), "v.csv", collapse=" ")))
v <- v %>% 
  mutate_at(vars(  landscape, sedclass, seddan, sedmil    ), factor)
v <- v %>% filter(complete.cases(.))

mtry <- read.csv(file.path(savepath, paste(vme, "minus", paste(substr(remove, 1,7), collapse=", "), "mtry.csv")))

selvar <- read.csv(file.path(savepath, paste(vme, "minus", paste(substr(remove, 1,7), collapse=", "), "selvar.csv")))
selvar <- pull(selvar)

all_names <- paste(#c(names(resp.dist0),
  selvar, collapse="+")

fmla0 <- as.formula(paste("vmeind_density ~ ", 
                          all_names))

```

# Load predictor file
```{r}
#load(file.path(savepath,"1_Predictors_out.RData"))
#pred.pix <- readRDS(file.path(savepath,"pred_data/pred.rds"))
pred.pix1 <- pred.pix[,selvar]
pred.pix1 <- subset(pred.pix1, complete.cases(pred.pix1@data))
pred.pix1@data <- pred.pix1@data %>% mutate_at(vars(  landscape, # need to precede this with a condition
                                                      sedclass,
                                                      seddan,
                                                      #sedmil
                                                      ), factor)
```

# Select model
```{r}
m <- readRDS(file = file.path(outpath, paste(paste(vme,"minus", paste(substr(remove, 1,7), collapse=", ")),
                                             paste0(name, ".rda"))))
m
```

# Using spatial pixels
```{r}
vme.rfd1 <- predict(m, cbind(
  #resp.dist0@data, 
  pred.pix1@data), 
                     type="quantiles", quantiles=quantiles)$predictions
## now more computational...
pred.pix1$vme_dens = vme.rfd1[,2]
## s.d. of the prediction error:
pred.pix1$vme_dens_var = (vme.rfd1[,3]-vme.rfd1[,1])/2
str(pred.pix1@data)
```

# visualize
```{r qrf-sic97-maps, echo=FALSE, fig.width=9, fig.cap="Predictions and prediction errors for the SIC97 data set."}
par(mfrow=c(1,2), oma=c(0,0,0,0.5), mar=c(0,0,1.5,1))
plot(raster(pred.pix1["vme_dens"]), col=leg, 
     main="Random Forest (RF)", axes=FALSE, box=FALSE)
#points(resp_spat, pch="+")
plot(raster(pred.pix1["vme_dens_var"]), col=rev(bpy.colors()), 
     main="Random Forest (RF) prediction error", axes=FALSE, box=FALSE)
#points(resp_spat, pch="+")
```

# save outputs
```{r}
writeGDAL(pred.pix1["vme_dens"],
          fname = file.path(file.path(outpath, name), paste(vme,"minus", paste(remove, collapse=", "), "pred.tif")),
          options = c("COMPRESS=DEFLATE"))

writeGDAL(pred.pix1["vme_dens_var"],
          fname = file.path(file.path(outpath, name), paste(vme,"minus", paste(remove, collapse=", "), "pred var.tif")),
          options = c("COMPRESS=DEFLATE"))

```
